Add Rawr on Blast and refactor iziswapExport (#9172)
* Add Rawr on Blast

* Remove unused code and refactor iziswapExport

---------
